updat

+++ server/service/admin/UserService.js
@@ -33,11 +33,28 @@ const UserService = {
     })
   },
   getList: async({id})=>{
+    console.log('-------------------------');
+    
+    console.log(id);
+    
+    // return id?UserModel.find({_id:id},['username','role','introduction','password'])
+    // :UserModel.find({},['username','role','avatar','introduction','gender'])
     return id?UserModel.find({_id:id},['username','role','introduction','password'])
-    :UserModel.find({},['username','role','avatar','introduction','gender'])
+    :UserModel.find()
+  },
+  findlist: async({username})=>{
+    console.log(username);
+    
+    return UserModel.find({username})
   },
   delList: async({_id}) =>{
-    return UserModel.deleteOne({_id})
+    let res = await UserModel.find({_id:_id},['username'])
+    // console.log(res.username);
+    if(res[0]._id == '66cc728c6329c503b3102e9e'){
+      return '你没有删除管理员的权限'
+    } else {
+      return UserModel.deleteOne({_id})
+    }
   },
   putList: async(body)=>{
     return UserModel.updateOne({_id:body._id},body)
